ascii-service/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.34.0
ascii_magic>=2.3.0
pillow>=12.0.0
python-multipart>=0.0.20
